fix: 리스트만 클라이언트로 분리
