+++ one-day-hero/src/app/mission/list/suggested/page.tsx
@@ -1,65 +1,16 @@
-"use client";
+import { redirect } from "next/navigation";
 
-import Link from "next/link";
-import { useRouter } from "next/navigation";
-import { useRef } from "react";
-
-import { getClientToken } from "@/app/utils/cookie";
-import Button from "@/components/common/Button";
-import Container from "@/components/common/Container";
-import MissionFullInfo from "@/components/common/Info/MissionFullInfo";
-import { useGetSuggestedMissionListFetch } from "@/services/missions";
+import { getServerToken } from "@/app/utils/auth";
+import SuggestedMissionList from "@/components/domain/mission/suggested/SuggestedMissionList";
 
 const SuggestedMissionPage = () => {
-  const token = getClientToken();
+  const token = getServerToken();
 
-  const router = useRouter();
-
-  if (!token) router.push("/login?redirect=");
-
-  const observerRef = useRef<HTMLDivElement | null>(null);
-
-  const { data } = useGetSuggestedMissionListFetch(
-    "1",
-    token ?? "",
-    observerRef
-  );
+  if (!token) redirect("/login?redirect=");
 
   return (
     <div className="mt-20 flex w-full max-w-screen-sm flex-col items-center justify-center space-y-4">
-      {data.map(({ id, mission }) => (
-        <Link
-          href={`/mission/${id}`}
-          className="flex w-full max-w-screen-sm justify-center"
-          key={id}>
-          <Container className="cs:w-11/12" missionStatus={mission.status}>
-            <MissionFullInfo
-              bookmarkCount={mission.bookmarkCount}
-              createdAt={mission.createdAt}
-              region={mission.region}
-              missionCategory={mission.missionCategory}
-              missionInfo={mission.missionInfo}
-            />
-            <div className="flex justify-center gap-1">
-              <Button
-                theme="cancel"
-                size="sm"
-                textSize="sm"
-                className="cs:h-10">
-                거절하기
-              </Button>
-              <Button
-                theme="primary"
-                size="sm"
-                textSize="sm"
-                className="cs:h-10">
-                채팅하기
-              </Button>
-            </div>
-          </Container>
-        </Link>
-      ))}
-      <div ref={observerRef} />
+      <SuggestedMissionList />
     </div>
   );
 };
